17/12/2018
correction bug deco au bout de 2 connexion

+++ src/action/ActionNePlusRester.java
@@ -32,7 +32,7 @@ public class ActionNePlusRester implements ActionListener{
 
 		this.vue.setContentPane(new panel_Connexion(this.vue, this.nb));
 		
-		if(Modele.resteConnecter()) {
+		if(Modele.resteConnecter() && this.nb == 0) {
 			Path path = Paths.get("stock.txt");
 			try {
 			    Files.delete(path);
